requirements.txt
python-dotenv
openai
streamlit
google-generativeai
anthropic
cohere
replicate
langchain
langchain-openai
langchain-google-genai
langchain_experimental
beautifulsoup4
faiss-cpu
pypdf
unstructured
networkx
openpyxl
rapidocr-onnxruntime
pandas